# backend/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RefuLearn API",
    "description": "A comprehensive API for the RefuLearn platform supporting refugee education and skill development",
    "version": "1.0.0",
    "contact": {
      "name": "RefuLearn Team",
      "email": "support@refulearn.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.refulearn.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    },
  "paths": {
...
    "/api/auth/users/{userId}": {
      "get": {
        "tags": ["auth"],
        "summary": "Get user by ID",
        "description": "Get a user by their ID (admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "User retrieved successfully" },
          "404": { "description": "User not found" }
        }
      },
      "put": {
        "tags": ["auth"],
        "summary": "Update user",
        "description": "Update a user by their ID (admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": { "type": "string" },
                  "lastName": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "role": { "type": "string", "enum": ["refugee", "instructor", "admin", "employer"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User updated successfully" },
          "404": { "description": "User not found" }
        }
      },
      "delete": {
        "tags": ["auth"],
        "summary": "Delete user",
        "description": "Delete a user by their ID (admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "User deleted successfully" },
          "404": { "description": "User not found" }
        }
      }
    },
...
    "/api/admin/users/{userId}": {
      "get": {
        "tags": ["admin"],
        "summary": "Get user by ID",
        "description": "Retrieve a specific user by their ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "User retrieved successfully" },
          "404": { "description": "User not found" }
        }
      },
      "put": {
        "tags": ["admin"],
        "summary": "Update user",
        "description": "Update a user's information",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": { "type": "string" },
                  "lastName": { "type": "string" },
                  "email": { "type": "string" },
                  "role": { "type": "string" },
                  "isActive": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User updated successfully" },
          "404": { "description": "User not found" }
        }
      },
      "delete": {
        "tags": ["admin"],
        "summary": "Delete user",
        "description": "Delete a user from the system",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "User deleted successfully" },
          "404": { "description": "User not found" }
        }
      }
    },
...
  }
}